Suppress warnings due to included Godot headers

+++ src/include/stratagus.h
@@ -131,7 +131,7 @@ extern void PrintOnStdErr(const char *format, ...);
 
 inline char *new_strdup(const char *str)
 {
-	int len = strlen(str) + 1;
+	int len = static_cast<int>(strlen(str) + 1);
 	char *newstr = new char[len];
 	strcpy_s(newstr, len, str);
 	return newstr;
